Implemented feedbacks on endpoints implementation

+++ app/controller/entryController.js
@@ -1,10 +1,10 @@
 import pool from '../db/pool';
 
 const badRequest = { status: 400, message: 'Bad Request' };
+
 const getEntry = (req, res) => {
-  pool.query('SELECT title, description FROM entries WHERE id = ($1)', [req.params.id], (err, dbRes) => {
+  pool.query('SELECT * FROM entries WHERE (id, user_id) = ($1, $2)', [req.params.id, req.userData.userId], (err, dbRes) => {
     if (err) {
-      // console.log(dbRes);
       const reply = { status: '500', message: 'Internal Server Error', description: 'Could not retrieve entry' };
       res.status(500).send(reply);
     } else {
@@ -15,20 +15,20 @@ const getEntry = (req, res) => {
       if (dbRes.rows[0] === undefined) {
         res.status(404).send(reply);
       } else {
-        res.status(200).send(dbRes.rows[0]);
+        const goodReply = { status: '200', message: 'Entry returned successfully', entry: dbRes.rows[0] };
+        res.status(200).send(goodReply);
       }
     }
   });
 };
 
 const getEntries = (req, res) => {
-  pool.query('SELECT * FROM entries ORDER BY id ASC', (err, dbRes) => {
+  pool.query('SELECT * FROM entries WHERE user_id = ($1) ORDER BY id DESC', [req.userData.userId], (err, dbRes) => {
     if (err) {
       // console.log(err.stack);
       const reply = { status: '500', message: 'Internal Server Error', description: 'Could not retrieve entry' };
       res.status(500).send(reply);
     } else {
-      // console.log(dbRes.rows[0]);
       const db = { entries: dbRes.rows, size: dbRes.rows.length };
       const reply = {
         status: '404', message: 'Entry Not Found',
@@ -36,7 +36,8 @@ const getEntries = (req, res) => {
       if (dbRes.rows === undefined) {
         res.status(404).send(reply);
       } else {
-        res.status(200).send(db);
+        const goodReply = { status: '200', message: 'Entries returned successfully', data: db };
+        res.status(200).send(goodReply);
       }
     }
   });
@@ -47,26 +48,24 @@ const createEntry = (req, res) => {
     badRequest.description = 'Body or title cannot be empty';
     res.status(400).send(badRequest);
   } else {
-    pool.query('INSERT INTO entries(title, description) values($1, $2)',
-      [req.body.title, req.body.text], (error) => {
+    pool.query('INSERT INTO entries(title, description, user_id) values($1, $2, $3)',
+      [req.body.title, req.body.text, req.userData.userId], (error) => {
         if (error) {
           const replyServer = { status: '500', message: 'Internal Server Error', description: 'Could not post entry' };
           res.status(500).send(replyServer);
         } else {
-          // res.status(200).send(reply);
-          pool.query('SELECT * FROM entries ORDER BY id ASC', (err, dbRes) => {
+          pool.query('SELECT * FROM entries WHERE user_id = ($1) ORDER BY id DESC', [req.userData.userId], (err, dbRes) => {
             if (err) {
-              // console.log(err.stack);
               const reply = { status: '500', message: 'Internal Server Error', description: 'Could not retrieve entry' };
               res.status(500).send(reply);
             } else {
-              // console.log(dbRes.rows[0]);
               const db = { entries: dbRes.rows, size: dbRes.rows.length };
               const reply = { status: '404', message: 'Entry Not Found' };
               if (dbRes.rows === undefined) {
                 res.status(404).send(reply);
               } else {
-                res.status(200).send(db);
+                const goodReply = { status: '201', message: 'Entry Created successfully', data: db };
+                res.status(201).send(goodReply);
               }
             }
           });
@@ -76,25 +75,24 @@ const createEntry = (req, res) => {
 };
 
 const updateEntry = (req, res) => {
-  pool.query('UPDATE entries SET title = ($1), description = ($2) WHERE id = ($3)',
-    [req.body.title, req.body.text, req.body.id], (error) => {
+  pool.query('UPDATE entries SET title = ($1), description = ($2) WHERE (id, user_id) = ($3, $4)',
+    [req.body.title, req.body.text, req.params.id, req.userData.userId], (error) => {
       if (error) {
         const replyServer = { status: '500', message: 'Internal Server Error', description: 'Could not update entry' };
         res.status(500).send(replyServer);
       } else {
-        pool.query('SELECT * FROM entries ORDER BY id ASC', (err, dbRes) => {
+        pool.query('SELECT * FROM entries WHERE user_id = ($1) ORDER BY id DESC', [req.userData.userId], (err, dbRes) => {
           if (err) {
-            // console.log(err.stack);
             const reply = { status: '500', message: 'Internal Server Error', description: 'Could not retrieve entry' };
             res.status(500).send(reply);
           } else {
-            // console.log(dbRes.rows[0]);
             const db = { entries: dbRes.rows, size: dbRes.rows.length };
             const reply = { status: '404', message: 'Entry Not Found' };
             if (dbRes.rows === undefined) {
               res.status(404).send(reply);
             } else {
-              res.status(200).send(db);
+              const goodReply = { status: '200', message: 'Entry Modified successfully', entry: db };
+              res.status(200).send(goodReply);
             }
           }
         });
@@ -104,13 +102,25 @@ const updateEntry = (req, res) => {
 
 const deleteEntry = (req, res) => {
   let reply = {};
-  pool.query('DELETE FROM entries WHERE id = ($1)', [req.params.id], (err) => {
+  pool.query('DELETE FROM entries WHERE (id, user_id) = ($1, $2)', [req.params.id, req.userData.userId], (err) => {
     if (err) {
       reply = { status: '500', message: 'Internal Server Error', description: 'Could not delete Entry' };
       res.status(500).send(reply);
     } else {
-      reply = { status: '200', message: 'Entry Deleted Successfully' };
-      res.status(200).send(reply);
+      pool.query('SELECT * FROM entries WHERE user_id = ($1) ORDER BY id DESC', [req.userData.userId], (error, dbRes) => {
+        if (error) {
+          res.status(500).send(reply);
+        } else {
+          const db = { entries: dbRes.rows, size: dbRes.rows.length };
+          const badReply = { status: '404', message: 'Entry Not Found' };
+          if (dbRes.rows === undefined) {
+            res.status(404).send(badReply);
+          } else {
+            const goodReply = { status: '200', message: 'Entry Deleted successfully', entry: db };
+            res.status(200).send(goodReply);
+          }
+        }
+      });
     }
   });
 };
